footer css,dashboard user, button delete and update mediacard

+++ frontend/src/pages/dashboard.js
@@ -11,10 +11,15 @@ export default function Dashboard() {
 
   const deleteUser = async (userId) => {
     try {
-      await axios.delete(`http://localhost:7000/tech/delete/${userId}`);
+      const confirmDelete = window.confirm('Are you sure you want to delete this user?');
+
+      if(confirmDelete){
+         await axios.delete(`http://localhost:7000/tech/delete/${userId}`);
       // Update the users state to reflect the deletion
       const updatedUsers = users.filter(user => user.id !== userId);
       setUsers(updatedUsers);
+      }
+     
     } catch (error) {
       console.error('Error deleting user:', error);
     }
@@ -47,7 +52,7 @@ export default function Dashboard() {
     // Inner function that calls the API
     const fetchUsers = async () => {
       try {
-       
+        
         const resp = await axios.get('http://localhost:7000/tech/getAllUsers'); 
 
        
@@ -76,12 +81,12 @@ export default function Dashboard() {
      <div id="dhashboard-filter">
         <div id="market-place-performance">Market Place Performance</div>
         <div id="select"> 
-            <select name="days" id="days">
+            {/* <select name="days" id="days">
 
         </select>
         <select name="all-categ" id="all-categ">
 
-        </select>
+        </select> */}
     </div>
        
      </div>
@@ -95,7 +100,7 @@ export default function Dashboard() {
             <span>Created at </span>
             <span>Updated at </span>
             <span>Price</span>
-            <span>Oweners</span>
+            <span>Description</span>
             <span>Color</span>
             </div>
